working proof of concept of git pack caching (in memory)
This is a very conservative cache; in addition to not caching refs checks, which is very much by design (since this software must never return a stale response), this current implementation has no understanding of git packs and so can only cache requests that are *identical*.

For example, a fresh clone will be cached against an earlier fresh clone, and a fetches from one client will be cached against a fetch from another client with the exact same existing commits, but the fetches will not cache against a full clone.

So there is a lot of scope for improvement, but this level of request caching is useful if you have a number of probably-identical clones such as a number of identical VMs all pulling the same repo.